added support for new symptoms (diarrhea, abdominal pain, anosmia and fatigue)
